+++ team.h
@@ -3,7 +3,7 @@
 
 struct team;
 
-struct team *team_create(char type, unsigned int number_of_antennas);
+struct team *team_create(char type, size_t number_of_antennas);
 void team_destroy(struct team *this_team);
 
 int team_add_device_sensor(struct team *this_team, unsigned int host_number, char *device_name, char *sensor_name);
